Separated messages to its own file
